+++ crates/stackable-telemetry/src/tracing/settings/mod.rs
@@ -85,7 +85,11 @@ impl SettingsBuilder {
     }
 
     /// Set specific [`FileLogSettings`].
-    pub fn file_log_settings_builder<P>(self, path: P) -> FileLogSettingsBuilder
+    pub fn file_log_settings_builder<P>(
+        self,
+        path: P,
+        filename_suffix: impl Into<String>,
+    ) -> FileLogSettingsBuilder
     where
         P: AsRef<Path>,
     {
@@ -93,6 +97,7 @@ impl SettingsBuilder {
             common_settings: self.build(),
             file_log_dir: path.as_ref().to_path_buf(),
             rotation_period: Rotation::NEVER,
+            filename_suffix: filename_suffix.into(),
         }
     }