+++ crates/stackable-telemetry/CHANGELOG.md
@@ -6,7 +6,7 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 
-- Allow customization of the rolling file appender [#995].
+- BREAKING: Allow customization of the rolling file appender [#995].
   - Add required `filename_suffix` field.
   - Add `with_rotation_period` method.
   - Add `with_max_log_files` method.