+++ crates/stackable-telemetry/CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+### Added
+
+- BREAKING: Allow customization of the rolling file appender [#995].
+  - Add required `filename_suffix` field.
+  - Add `with_rotation_period` method.
+  - Add `with_max_log_files` method.
+
+[#995]: https://github.com/stackabletech/operator-rs/pull/995
+
 ## [0.3.0] - 2025-01-30
 
 ### Added